server/requirements.txt
# Usage: pip3 install -r requirements.txt

# Flask
flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.5
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.5.1

# Misc
alembic==1.7.7
pylint==2.13.5
coverage==6.3.2
sqlalchemy==1.4.35
SQLAlchemy-Utils==0.38.2
blueprint==3.4.2
uwsgi==2.0.20
psycopg2==2.9.3
Werkzeug==2.1.1
marshmallow==3.15.0
PyJWT==2.3.0
requests==2.27.1

# Machine Learning
river==0.10.1
scikit-learn==1.0.2
